tried to get collisions working but having issues
I think the issues lie in the collision nodes being inside of the skelton3d node. i think that might be why i am having issues. i wonder if i can make a rigid body node that can be instantiated in the main scene but tween to the skeleton 3d nodes.